+++ packages/components/CHANGELOG.md
@@ -3,6 +3,82 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.19.2](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.19.1...@jobber/components@4.19.2) (2023-06-01)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.19.1](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.19.0...@jobber/components@4.19.1) (2023-06-01)
+
+
+### Bug Fixes
+
+* **components:** TECH-22 Fixing Grid prop type ([#1298](https://github.com/GetJobber/atlantis/issues/1298)) ([551fd33](https://github.com/GetJobber/atlantis/commit/551fd3314c95286505f3c5e6cdc568e0a50f9587))
+
+
+
+
+
+# [4.19.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.18.9...@jobber/components@4.19.0) (2023-05-31)
+
+
+### Features
+
+* **components:** Add Controller to Checkbox ([#1296](https://github.com/GetJobber/atlantis/issues/1296)) ([91fafd6](https://github.com/GetJobber/atlantis/commit/91fafd6b15c307f002e86540b1c1c75a99389888))
+
+
+
+
+
+## [4.18.9](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.18.8...@jobber/components@4.18.9) (2023-05-29)
+
+
+### Bug Fixes
+
+* **components:** adjust grid and page spacing ([#1287](https://github.com/GetJobber/atlantis/issues/1287)) ([3091a6c](https://github.com/GetJobber/atlantis/commit/3091a6ccf98b2dc1b11137a387a9d32121068514))
+
+
+
+
+
+## [4.18.8](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.18.7...@jobber/components@4.18.8) (2023-05-29)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.18.7](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.18.6...@jobber/components@4.18.7) (2023-05-29)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.18.6](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.18.5...@jobber/components@4.18.6) (2023-05-26)
+
+
+### Bug Fixes
+
+* enable prefix and suffix on InputNumber ([#1289](https://github.com/GetJobber/atlantis/issues/1289)) ([ff02f7a](https://github.com/GetJobber/atlantis/commit/ff02f7ad2805d6473176fe1fe92064a330a4d2cb))
+
+
+
+
+
+## [4.18.5](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.18.4...@jobber/components@4.18.5) (2023-05-26)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
 ## [4.18.4](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.18.3...@jobber/components@4.18.4) (2023-05-26)